pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aoc"
version = "0.0.1"
authors = [
  { name="Emil Kloeden", email="emilkloeden@gmail.com" },
]
description = "Advent of Code scaffolding CLI"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-dotenv == 0.21.0",
    "requests == 2.28.1"
]


[project.urls]
"Homepage" = "https://github.com/emilkloeden/aoc"
"Bug Tracker" = "https://github.com/emilkloeden/aoc/issues"



[project.scripts]
aoc = "aoc.__main__:main"